+++ app/Controller/CoInvitesController.php
@@ -118,6 +118,22 @@ class CoInvitesController extends AppController {
    */
   
   protected function calculateImpliedCoId($data = null) {
+    if($this->action == "add" && !empty($data['co_person_id'])) {
+      // For historical reasons, CO ID is also passed in, but we need to ignore
+      // it and lookup via the CO Person ID.
+
+      $coId = $this->CoInvite->CoPerson->field('co_id',
+                                               array('id' => $data['co_person_id']));
+
+      if($coId) {
+        return $coId;
+      } else {
+        throw new InvalidArgumentException(_txt('er.notfound',
+                                                array(_txt('ct.co_people.1'),
+                                                      filter_var($data['co_person_id'],FILTER_SANITIZE_SPECIAL_CHARS))));
+      }
+    }
+
     if($this->action == "confirm"
        || $this->action == "authconfirm"
        || $this->action == "reply") {
@@ -155,7 +171,24 @@ class CoInvitesController extends AppController {
                                                       filter_var($this->request->params['named']['copersonid'],FILTER_SANITIZE_SPECIAL_CHARS))));
       }
     }
-    
+
+    if($this->action === "verifyEmailAddress" && !empty($this->request->params['named']['email_address_id'])) {
+      // Get CO Person ID from the Email entry
+      $this->EmailAddress->id = $this->request->params['named']['email_address_id'];
+      $co_person_id = $this->EmailAddress->field('co_person_id');
+
+      if($co_person_id) {
+        $coId = $this->EmailAddress->CoPerson->field('co_id', array('id' => $co_person_id));
+        if($coId) {
+          return $coId;
+        } else {
+          throw new InvalidArgumentException(_txt('er.notfound',
+                                                  array(_txt('ct.email_addresses.1'),
+                                                        filter_var($this->request->params['named']['email_address_id'],FILTER_SANITIZE_SPECIAL_CHARS))));
+        }
+      }
+    }
+
     return null;
   }
   
@@ -237,26 +270,45 @@ class CoInvitesController extends AppController {
     $roles = $this->Role->calculateCMRoles();
     
     $managed = false;
-    
+    $self = false;
+    $canRequestVerification = false;
+
+    // $managed
     if(!empty($roles['copersonid'])
        && $this->action == 'send'
        && !empty($this->request->params['named']['copersonid'])) {
       $managed = $this->Role->isCoOrCouAdminForCoPerson($roles['copersonid'],
                                                         $this->request->params['named']['copersonid']);
     }
-    
+
+    if(!empty($this->request->params['named']['email_address_id'])
+       && !empty($roles['copersonid'])) {
+      // $self
+      // Get CO Person ID from the Email entry
+      $this->EmailAddress->id = $this->request->params['named']['email_address_id'];
+      $co_person_id = $this->EmailAddress->field('co_person_id');
+      // $this->Session->read('Auth.User.co_person_id') written in the session by calculateCMRoles()
+      if ($co_person_id === $roles['copersonid']) {
+        $self = true;
+      }
+
+      // $canRequestVerification
+      $canRequestVerification = $this->Role->canRequestVerificationOfEmailAddress($roles['copersonid'], $this->request->params['named']['email_address_id']);
+    }
+
     // Construct the permission set for this user, which will also be passed to the view.
     $p = array();
     
     // Determine what operations this user can perform
     
     // Send an invite? (REST only)
-    $p['add'] = $roles['apiuser'];
+    $p['add'] = $roles['apiuser'] && ($roles['cmadmin'] || $roles['coadmin']);
     
     // Confirm an invite? (HTML, auth required)
     $p['authconfirm'] = true;
     
     // Verify an email address?
+    // We basically do the same thing as authconfirm().
     $p['authverify'] = true;
 
     // Confirm an invite? (HTML only)
@@ -266,25 +318,23 @@ class CoInvitesController extends AppController {
     $p['decline'] = true;
     
     // Confirm or decline an invite? (REST only)
-    $p['index'] = $roles['apiuser'];
+    $p['index'] = $roles['apiuser'] && ($roles['cmadmin'] || $roles['coadmin']);
     
     // Reply to an invite? (HTML only)
     $p['reply'] = true;
     
     // Send an invite? (HTML only)
     
-    $p['send'] = ($roles['cmadmin']
-                  || ($managed && ($roles['coadmin'] || $roles['couadmin'])));
+    $p['send'] = $roles['cmadmin']
+                 || $roles['coadmin']
+                 || ($managed && $roles['couadmin']);
     
     // Request verification of an email address?
     // This needs to correlate with EmailAddressesController
-    $p['verifyEmailAddress'] = (!empty($this->request->params['named']['email_address_id'])
-                                ? ($roles['cmadmin']
-                                   ||
-                                   $this->Role->canRequestVerificationOfEmailAddress($roles['copersonid'],
-                                                                                     $this->request->params['named']['email_address_id']))
-                                : false);
-    
+    $p['verifyEmailAddress'] = $roles['cmadmin']             // I am the Platform Administrator
+                               || $self                      // It is my Email Address linked to my Canvas
+                               || $canRequestVerification;   // I have been given authorization to request the verification of the Email
+
     $this->set('permissions', $p);
     return $p[$this->action];
   }
@@ -808,7 +858,7 @@ class CoInvitesController extends AppController {
                                     $this->Session->read('Auth.User.co_person_id'),
                                     $ea['EmailAddress']['mail'],
                                     null, // use default from address
-                                    (!empty($this->cur_co['Co']['name']) ? $this->cur_co['Co']['name'] : _txt('er.unknown')),
+                                    (!empty($this->cur_co['Co']['name']) ? $this->cur_co['Co']['name'] : _txt('er.unknown.def')),
                                     _txt('em.invite.subject.ver'),
                                     _txt('em.invite.body.ver'),
                                     $ea['EmailAddress']['id']);