+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - CoPersonRole Status Fix during Petition
 - Expiration Policy have to implement Changelog Behavior
 - lang.php localization do not apply to Job Shells
+- Enable Email Verification by unpriviledged users
 
 ### Changed